fix: leave astroport stableswap as original. move all tests inside helpers files
