+++ src/implementations/astroport/mod.rs
@@ -6,6 +6,3 @@ mod staking;
 
 pub use pool::AstroportPool;
 pub use staking::AstroportStaking;
-
-#[cfg(test)]
-mod helper_test;
