+++ src/implementations/junoswap/mod.rs
@@ -6,6 +6,3 @@ mod staking;
 
 pub use pool::*;
 pub use staking::*;
-
-#[cfg(test)]
-mod helper_test;
